% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nis_exc_dx.r
\name{nis_exc_dx}
\alias{nis_exc_dx}
\title{NIS Exclude by Diagnostic Codes}
\usage{
nis_exc_dx(df, dx_codes)
}
\arguments{
\item{dx_codes}{List of ICD9 diagnostic codes given as strings}
}
\description{
Returns rows from NIS Core that do not have the diagnostic codes
}
\examples{
nis_exc_dx(df, dx_codes)
}
\keyword{nis}

