% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nis_exc_pr.r
\name{nis_exc_pr}
\alias{nis_exc_pr}
\title{NIS Exclude by Procedure Codes}
\usage{
nis_exc_pr(df, pr_codes)
}
\arguments{
\item{pr_codes}{List of ICD9 procedure codes given as strings}
}
\description{
Returns rows from NIS Core that do not have the procedure codes
}
\examples{
nis_exc_pr(df, pr_codes)
}
\keyword{nis}

