% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nis_inc_dx.r
\name{nis_inc_dx}
\alias{nis_inc_dx}
\title{NIS Include by Diagnostic Codes}
\usage{
nis_inc_dx(df, dx_codes)
}
\arguments{
\item{dx_codes}{List of ICD9 diagnostic codes given as strings}
}
\description{
Returns rows from NIS Core that have the diagnostic codes
}
\examples{
nis_inc_dx(df, dx_codes)
}
\keyword{nis}

