% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nis_inc_pr.r
\name{nis_inc_pr}
\alias{nis_inc_pr}
\title{NIS Include by Procedure Codes}
\usage{
nis_inc_pr(df, pr_codes)
}
\arguments{
\item{pr_codes}{List of ICD9 procedure codes given as strings}
}
\description{
Returns rows from NIS Core that have the procedure codes
}
\examples{
nis_inc_pr(df, pr_codes)
}
\keyword{nis}

