% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nis_sql_dx.r
\name{nis_sql_dx}
\alias{nis_sql_dx}
\title{Query NIS SQLite Core and Severity Files by Year and Diagnostic ICD9 Codes}
\usage{
nis_sql_dx(year, dx_codes)
}
\arguments{
\item{year}{NIS Year}

\item{dx_codes}{List of ICD9 diagnoses given as strings}
}
\description{
Query NIS by year and ICD9 codes and returns a Dataframe of the Core and Severity data
}
\examples{
nis_sql_dx(year, dx_codes)
}
\keyword{nis}

