% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nis_sql_ecode.r
\name{nis_sql_ecode}
\alias{nis_sql_ecode}
\title{Query NIS SQLite Core and Severity Files by Year and ICD9 E Codes}
\usage{
nis_sql_ecode(year, nis_path = "~/NIS", ecodes)
}
\arguments{
\item{year}{NIS Year}

\item{nis_path}{Path to the NIS DB files. Defaults to "~/NIS"}

\item{ecodes}{List of ICD9 procedures given as strings}
}
\description{
Query NIS by year and ICD9 E Codes and returns a Dataframe of the Core and Severity data
}
\examples{
nis_sql_ecode(year, ecodes)
}
\keyword{nis}

