% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nis_sql_pr.r
\name{nis_sql_pr}
\alias{nis_sql_pr}
\title{Query NIS SQLite Core and Severity Files by Year and Procedural ICD9 Codes}
\usage{
nis_sql_pr(year, nis_path = "~/NIS", pr_codes)
}
\arguments{
\item{year}{NIS Year}

\item{nis_path}{Path to the NIS DB files. Defaults to "~/NIS"}

\item{pr_codes}{List of ICD9 procedures given as strings}
}
\description{
Query NIS by year and ICD9 codes and returns a Dataframe of the Core and Severity data
}
\examples{
nis_sql_pr(year, pr_codes)
}
\keyword{nis}

